pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nulltracer"
version = "0.9.0"
description = "GPU-accelerated ray tracing through Kerr-Newman spacetime"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ethan Knox", email = "ethank5149@gmail.com"},
]
keywords = [
    "black-hole",
    "ray-tracing",
    "cuda",
    "general-relativity",
    "gpu",
    "scientific-computing",
    "kerr-metric",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "scipy>=1.10",
    "numpy>=1.24",
    "cupy-cuda12x>=13.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
notebook = [
    "jupyterlab>=4.0",
    "ipywidgets>=8.0",
]
server = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "Pillow>=10.0",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/ethank5149/nulltracer"
Repository = "https://github.com/ethank5149/nulltracer"

[tool.setuptools.packages.find]
include = ["nulltracer*"]

[tool.setuptools.package-data]
nulltracer = ["kernels/*.cu", "kernels/integrators/*.cu"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 120
select = ["E", "F", "W"]
ignore = ["E501"]